% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideHazard.R
\name{distance_to_fault}
\alias{distance_to_fault}
\title{Distancia a fallas geologicas}
\usage{
distance_to_fault(dem, faults)
}
\arguments{
\item{dem}{Un objeto SpatRaster que sirve como referencia espacial}

\item{faults}{Un objeto sf con las diferentes fallas poligonizadas}
}
\value{
Un objeto SpatRaster con valores de distancia
}
\description{
Calcula la distancia Euclidiana de cada celda del raster a la linea de falla
mas proxima
}
