% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideHazard.R
\name{hazard_index}
\alias{hazard_index}
\title{Landslide Hazard Index}
\usage{
hazard_index(slope, distance, w_slope = 0.5, w_distance = 0.5)
}
\arguments{
\item{slope}{SpatRaster de pendiente (normalizado 0–1) heredado de la funcion
slope_func.}

\item{distance}{SpatRaster de distancia a fallas (normalizado 0–1) heredado
de la funcion distance_to_fault.}

\item{w_slope}{Peso de la pendiente (default = 0.5).}

\item{w_distance}{Peso de la distancia (default = 0.5).}
}
\value{
SpatRaster con índice de peligrosidad.
}
\description{
Calcula un índice simple de peligrosidad por deslizamiento
combinando pendiente y distancia a fallas.
}
\examples{
library(terra)

# Crear raster simulado
r <- rast(nrows = 50, ncols = 50,
          xmin = 0, xmax = 1000,
          ymin = 0, ymax = 1000,
          crs = "EPSG:25830")

values(r) <- runif(ncell(r), 0, 100)

slope_n <- normalize_minmax(r)
dist_n  <- normalize_minmax(r)

hazard <- hazard_index(slope_n, dist_n,
                            w_slope = 0.6,
                            w_distance = 0.4)

plot(hazard)
}
