% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideHazard.R
\name{normalize_minmax}
\alias{normalize_minmax}
\title{Normalize raster (min-max)}
\usage{
normalize_minmax(x)
}
\arguments{
\item{x}{Objetos SpatRaster heredados de las funciones distance_to_fault y
slope_func}
}
\value{
SpatRaster normalizado entre 0 y 1
}
\description{
Normaliza un SpatRaster usando el método min-max.
}
\examples{
library(terra)

r <- rast(nrows = 10, ncols = 10,
          xmin = 0, xmax = 100,
          ymin = 0, ymax = 100,
          crs = "EPSG:25830")

values(r) <- runif(ncell(r), 10, 50)

r_norm <- normalize_minmax(r)
terra::global(r_norm, c("min", "max"))
}
