% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideHazard.R
\name{slope_func}
\alias{slope_func}
\title{Calculate slope from a DEM}
\usage{
slope_func(dem)
}
\arguments{
\item{dem}{Un objeto SpatRaster que represente el Modelo Digital de Elevacion
(DEM).}
}
\value{
Un objeto SpatRaster con valores de pendientes expresados en unidades
de grados.
}
\description{
Calcula la pendiente de un Modelo Digital de Elevacion(DEM) en grados.
}
\examples{
library(terra)

dem <- rast(nrows = 50, ncols = 50,
            xmin = 0, xmax = 1000,
            ymin = 0, ymax = 1000,
            crs = "EPSG:25830")

values(dem) <- runif(ncell(dem), 0, 100)

slope <- slope_func(dem)
plot(slope)

}
